Remove repetitive `self.` in class method declarations
We have three classes/modules that contain no instance methods, only
class methods (and the fact that one of these is a class for which
instantiation is meaningless suggests that it should be a module; I'll
deal with that in a separate commit).

As they are just containers for class methods, get rid of the noise of
the repetitive `self.` in the method declarations by declaring the
methods in the module's metaclasses.