Marginally simplify prefix_len handling
Instead of adding 1 to the prefix handling in multiple places, do it
once only, when initially setting it.

This uncovered a bug in the prefix handling in the `GitFileScanner`,
where we are passing in 0 as a `prefix_len` to the `#path_excluded?`
method, and it was getting bumped to 1, trimming off the first character
in the path and making it pretty much useless.

In testing this I noticed that the `'wildignore'` behavior is
fundamentally different between the `RubyFileScanner` and the other two
scanners that potentially consult `'wildignore'` (the `GitFileScanner`
and the `FindFileScanner`).

The Ruby scanner uses DFS and immediately terminates the recursion on
seeing an ignored path. As a result, if I ignore "node_modules" then I
won't see any files under that directory.

With the other scanners, however, such an ignore would only prevent me
from seeing a file called "node_modules", but not any file rooted
underneath a directory with that name (eg. "node_modules/project/foo").

I hadn't noticed this before because I only ever really use the
`RubyFileScanner` (for small projects) and the `WatchmanFileScanner`
(for large projects).