doc: Document file scanner 'wildignore' differences
Perhaps I knew about these differences once and forgot about it; I just
noticed in preparing the last commit that the "ruby" scanner is not like
"git" or "find".

Document the differences, and update the examples that were using ".git"
and ".hg" (which work as you'd want only with the "ruby" scanner) to
instead use "**/.git/*" and "**/.hg/*" (which will work for all the
scanners except the "watchman" scanner, which doesn't use 'wildignore'
at all).