Bail earlier for ineligible dot-file matches
This gets us a nice speed-up for the typical workloads (ie. everything
but the atypcial pathological test case); below are the best of three
runs before and after this change:

Before:

                          user     system      total        real
  pathological        0.110000   0.000000   0.110000 (  0.116645)
  command-t           0.360000   0.000000   0.360000 (  0.357542)
  chromium (subset)   0.790000   0.010000   0.800000 (  0.383790)
  chromium (whole)    3.290000   0.010000   3.300000 (  1.288277)

After:

                          user     system      total        real
  pathological        0.110000   0.000000   0.110000 (  0.117410)
  command-t           0.400000   0.000000   0.400000 (  0.403234)
  chromium (subset)   0.560000   0.010000   0.570000 (  0.310002)
  chromium (whole)    2.460000   0.010000   2.470000 (  1.053952)

So we knock about 19% off the chromium (subset) time, and about 18% off
the chromium (whole) time. I am pretty surprised at the size of the
gains here, given the small number of dot-files in the benchmark test
data, but I added some specs to increase my confidence and they suggest
that the behavior hasn't changed, so let's run with this and see what
happens.

The code here is mostly simpler, because the dot-file logic is better
localized. The conditionals themselves continue to be a bit gnarly/ugly.