Avoid hang on Windows with g:CommandTTraverseSCM != "pwd"
The `while` loop in `PathUtilties#nearest_ancestor` was never
terminating on Windows because the root of the filesystem is not "/" but
rather "C:/" (or some similar drive letter).

The original fix implemented here was to replace:

  return nil if path == '/'

with:

  return nil if path == '/' || path =~ %r{\A[A-Z]:[/\\]\z}

This looks for any letter, followed by a colon, followed by "/" or "\".
In my testing on a single Windows VM (running XP and Ruby 1.9) I found
that "/" is used, but I wanted to be a bit more liberal in what I accept
here in order to minimize the risk of running into an infinite loop on
other versions of Windows, or other versions of Ruby.

In the end, however, I went with the more future-proof fix of checking
for whether the loop has advanced.

Fixes:

https://github.com/wincent/command-t/issues/110