Handle 'hidden' unwanted split edge case
If you had `'hidden'` off and a modified buffer in two windows, the old
logic here would run `:sp` instead of the `:e`, causing an unwanted
split.

We used `:sp` in order to avoid seeing an error message from Vim, but it
turns out that in this case Vim won't actually show an error because it
considers the extra window containing the modified buffer to mean that
you're not actually abandoning the buffer with unsaved changes.

The fix here detects that scenario by counting windows with the current
buffer in them. Note that this won't consider windows in other tabs,
because Vim doesn't expose those at all through `VIM::Window`.

Fixes:

https://github.com/wincent/command-t/issues/124