Move CommandT::VIM::PathUtilities out of VIM namespace
Other than a reference to `VIM::pwd`, there's nothing in this module
that makes it belong in the `VIM` namespace (we should reserve that for
extensions to Vim's Ruby runtime API).

So, this commit moves it up a level and updates the `require` statements
(plus adds one missing one, and removes one unnecessary one).