Replace and/or with &&/||
As a matter of hygiene, `and` and `or` are best avoided in typical
conditional contexts (their lower precedence can catch the unwary by
surprise). About the only place I feel they are called for now are
exceedingly rare post-fix statements (eg. `do_something and raise 'foo'`
and the like).

So, replace them with with `&&` and `||`. In addition to providing a
friendlier precedence, these are actually more readable, despite their
superficial ugliness (it's easier for the brain to match symbols amidst
a larger expression than English words among a string of words).

Additionally, change one `while ... do` to `while ...` for consistency;
the `do` is permitted here but best reserved for block (lambda) syntax.