Fix truthiness check when setting default SCM markers
I've been writing too much JavaScript lately, so stupidly expected
`markers && markers.length` to check for the presence and non-emptiness
of the `markers` array, which won't work because `0` is truthy in Ruby.

Switch to `Enumerable#any?` instead, which is roughly equivalent.
(Roughly, because `[nil, false].any?` is `false`, even though the array
has a non-zero length; the difference here shouldn't matter though.)

Thanks to Andri Möll for pointing out the mistake.