doc: CommandT needs to be built and run with the version of Ruby same as Vim
If you are an RVM or rbenv user, CommandT has to be built and run with
the same version of Ruby that Vim is linked against. Otherwise CommandT
won't work. So you either have to set that version of Ruby as default or
you need to re-build Vim with a version of Ruby you prefer.