Teach get_cursor_highlight to return an array
As of cfd65e2 `#get_cursor_highlight` may actually return more than one
`:highlight` command in order to support restoring highlights that
involve patterns like:

  :highlight! link Cursor VisualNOS

Rather than returning a string with another command tacked on, let's
just always return an array.

The one possible gotcha here is that this kind of rule must be set up
using `:highlight!` and not `:highlight`, or Vim will complain with:

  E414: group has settings, highlight link ignored

So, this commit changes the `#show_cursor` to always use `:highlight!`
instead of `:highlight`. It's possible this might have unintended
side-effects, so I'm going to follow up with a tweak that makes it a
little more nuanced.