Only use `highlight!` when needed
As mentioned in ab15fa2831b, we previously used `:highlight` for
everything, but the need to use `:highlight!` for some specific cases
led us to use `:highlight!` for everything.

This commit makes things a little more nuanced by marking the specific
cases that need the bang, and using it only in those cases.