Gather metadata and prepare to use it for installation diagnostics
When you run `extconf.rb` we get the compile-time `RUBY_VERSION` and
`RUBY_PATCHLEVEL` so that we can show it later on in a diagnostic
message if you try to run in a version of Vim with a mismatched Ruby.

Note that I say "prepare" to use it because in looking over this, I've
realized that the existing `Stub` module is effectively useless. We're
no longer eagerly loading the extension, which means that a failure to
compile will never be the cause of us using `CommandT::Stub`. In fact,
in order to get to the stub line, some other basic `require` has to have
failed, which in turn means things are busted and the `require` for
`Stub` likely won't work either.

I'll address that in a separate commit.