Eagerly load extension, to detect compilation problems earlier
As noted in the previous commit, the logic here wasn't really doing much
because we were only lazily loading the extension, so the likely
codepath was:

1. Try to load controller, fail
2. Modify load path
3. Try to load controller, succeed

The comment in there about falling back to the stub if there was a
problem with C extension wasn't really accurate. For example, if you
forgot to compile the extension, you'd still go through the sequence
above because we weren't even attempting to require it. Similarly, if
you compiled with the wrong Ruby version, we wouldn't find out about it
until later (and likely, with a segfault).

So, this commit reworks the logic here to do two new things:

- explicitly check run-time Ruby version against compile-time version;
  if we don't have metadata about the compile-time Ruby version we skip
  the check
- eagerly load the extension so that we can report problems via the
  `Stub` as we originally intended; this will catch the case where the
  user forgets to compile

Note that there is still one case that we can't handle, and that is the
one where the extension was previously compiled with the wrong version
of Ruby, but with an older version of Command-T so no metadata.rb file
existed or was updated. In that scenario, we skip the version check and
try to load the extension; because it exists but is the wrong version,
we'll probably segfault. From experience, segfaults are pretty much
assured if the mismatch is large (eg. 1.8.7 vs 2.0.0). If it's just a
difference in patchlevel then you _might_ get a segfault at some later
time (like during a search).

In any case, I think this commit leaves us in a better place. I tested
it with the following combinations:

- mismatched `metadata.rb`: prints diagnostic info
- unknown/default `metadata.rb`: loads and runs
- missing extension: prints diagnostic info
- unknown/default `metadata.rb` + big version mismatch (1.8.7 vs 2.0.0):
  segfaults