Clear CommandTMRUBuffer augroup before adding to it
It's true that this entire file is protected by a guard, but in case
somebody really, really wants to re-evaluate it they could use `:unlet`
to circumvent the guard and then re-source the file.

So, let's just follow standard best practice of clearing the entire
augroup before adding to it.