Avoid W10 when running in read-only mode
This was originally reported here:

https://wincent.com/issues/2088

But for some reason it didn't repro for me then, but it definitely does
now, with both `vim -R` and `view`. As soon as we try to make an
assignment to `@@buffer[...]` Vim issues the warning.

Make sure this doesn't happen by explicitly setting `noreadonly` in the
"GoToFile" buffer.