Fix infinite loop on <tab>
Whoops; a typo meant we were wrapping `#toggle_focus` in a guard, and
then wrapping the wrapped version in another guard, which meant that it
would end up calling itself repeatedly. Ouch.

The added regression spec isn't super awesome; it passes with the fix,
but hangs without the fix, so it is more documentation of what went
wrong than a real regression test per-se.

Fixes:

  https://github.com/wincent/command-t/issues/157