Fix release note preparation bug
I broke this in bc379eddf9. Ironically, at the time I said:

> As a matter of hygiene, `and` and `or` are best avoided in typical
> conditional contexts (their lower precedence can catch the unwary by
> surprise).

I still think that `and` can be surprising due to precedence, but here I
got stung by another arguable smell: using an assignment in a `while`
condition. It's a common pattern, and in C GCC will warn you unless you
write it as `((foo = thing))`, but in Ruby there's a strong idiomatic
tendency to avoid parens. I've met in the middle here by adding single
parens but not double ones.