Make statusline more useful
I don't want to actually set the statusline, because that might break
people's existing set-ups (they may be doing some special statusline
set-up of their own), but I do want to make it easier to customize the
statusline. So, this commit does two things:

(1) Sets a "command-t" `'filetype'`, which can be used to identify the
    Command-T match listing window (eg. from an `autocmd`) and take
    action on it, such as setting the statusline.

(2) Instead of just setting `GoToFile` at the buffer name, make it
    something more meaningful, like:

      Command-T [Buffers]
      Command-T [Files]
      Command-T [Jumps]
      Command-T [Recent]
      Command-T [Tags]

Note that I'll have to do a major version bump when this goes out,
because people may be relying on the old buffer name, as it was
previously the only way to identify Command-T.

I'll make a subsequent commit to my dotfiles repo showing how to
leverage this functionality, and then I'll update the docs here as well.