Mark a bunch of internal functions as "private"
These aren't strictly "private" in the way that script-local functions
would be, but moving them into the "private" autoloaded namespace makes
it obvious that they are not intended for external use, at least.

In order to make them actually inaccessible, I'd have to re-write a
large chunk of Command-T in Vimscript so that I could use script-local
functions, but that doesn't sound like much fun.