Switch from vimball to zip for distribution
Not a fan of the vimball format, so the move to 2.0 seems like as good a
time as any to get rid of it.

This commit is almost entirely untested, although I did run `rake
archive` at least. I guess I'll get around to testing the full flow when
it comes time for a release...