Future-proof copyright year ranges
I used to like manually updating these whenever a particular file was
touched because it gave me an indication at a glance of how "fresh" a
file was (ie. I could see files which hadn't been touched in years).

But now, laziness is starting to weigh more heavily than curiosity, so I
am updating these to be future-proof, once and for all.