Show deprecation warnings for deleted functions
Given that most people use Command-T via a plug-in manages they end up
tracking `HEAD` of `master` and there is no nice way to warn about
upcoming breaking changes (ie. there are no "releases" for these people
and release notes are meaningless; furthermore, people may end up
upgrading without any conscious decision to do so).

So, add functions that show an error when a deprecated function is
called, and then call through to the new function.

Note that depending on the function being called, the screen may be
cleared which means that the user may not see the error. I might add a
prompt here to address this issue because many users may not be aware of
the `:messages` command which can be used to display previous error
messages if they scroll or otherwise move out of view.