Avoid unwanted scrolling when showing/hiding match window
Thanks to @jordwalke for pointing out this problem and sharing a gist
showing that the unwanted displacement can be avoided by moving the
cursor in each window away from the side on which a split is going to be
created.

This is pretty much a ghastly hack, so I'll want to bake it for a long
time to be sure that it's worthy of landing on master. Note the
unpleasant Vim APIs that make it impossible to get useful information
from the windows without switching to them first. This requires us to do
some awkward save/restore shenanigans and a bunch of manual focus
management.