Fix spurious test failure
Commit 8e15beb added a new `g:CommandTRecursiveMatch` setting, which
defaults to "true" in the context of Command-T running inside Vim (ie.
preserving the previous behavior) but it does not default to "true" in
the artificial context of the test suite, causing one of the specs to
fail spuriously.

Fix that by explicitly setting `:recurse => true` in the test suite.

(Of course, maybe it _should_ be defaulting to true everywhere, for
consistency, but I wanted to keep the fix minimal before looking into a
deeper/riskier refactoring.)

Fixes: https://github.com/wincent/command-t/issues/189