Update Watchman binary protocol documentation URL
I noticed that this URL was 404ing, so I tracked down where I think it
moved to and updated the URL here. I considered referencing a commit
hash instead of master to prevent this from happening again but decided
against it because it will potentially be sending folks to outdated
documentation.