Avoid unnecessary recursing/scoring when string can't match
Borrow one of the optimizations to the Command-T scoring algorithm
from:

https://github.com/hansonw/fuzzy-native

Basically, we do a pre-scan to check whether the needle can possibly
match in the haystack, and if not, don't even bother trying to score it.

Additionally, we record the rightmost seen index in the haystack for
each character in the needle; we'll be using that later to prune the
search space during scoring, but for now I wanted to get this
optimization in on its own so I can benchmark its effect in isolation.

  Before (best of three runs):

                            user     system      total        real
    pathological        0.100000   0.000000   0.100000 (  0.091369)
    command-t           0.270000   0.000000   0.270000 (  0.267367)
    chromium (subset)   0.360000   0.020000   0.380000 (  0.157142)
    chromium (whole)    1.640000   0.000000   1.640000 (  0.562000)

  After (best of three runs):

                            user     system      total        real
    pathological        0.100000   0.000000   0.100000 (  0.096277)
    command-t           0.230000   0.000000   0.230000 (  0.232563)
    chromium (subset)   0.220000   0.020000   0.240000 (  0.141075)
    chromium (whole)    0.700000   0.000000   0.700000 (  0.439138)

Big improvements across the board; the most impressive delta is in the
large (chromium) dataset, where (user) CPU time dropped from 1.64
seconds to 0.7 seconds (system CPU time is negligible). On this
multicore machine, that corresponded to a still-impressive drop from
0.56 to 0.44 seconds of wall clock time (a 21% reduction).