Refactor scoring for clarity
Perf appears to be unaffected by this (which is good), but I think the
result is much clearer. Changes:

- Replace nested `for` loops + recursion with a single `for` loop plus
  recursion; the amount of work being done should be about the same as
  we're still recursing once for each matched character, and we're still
  early-returning in the same places thanks to memoization.
- Simplify calling convention of `recursive_match`: we no longer pass in
  `last_idx` (instead we derive it inside the function) nor `score`.
- No more `goto`.
- Return score directly from `calculate_match` instead of using
  cumbersome `out` parameter.
- Use some macros to name things for clarity (`NON_MATCH`, `UNSET`).
- Style tweaks.

The core idea here (of replacing one of the `for` loops with recursion)
was stolen from https://github.com/hansonw/fuzzy-native -- which itself
is a much clearer version of the old Command-T algorithm here, that had
grown organically over time. This makes the dynamic programming nature
of the approach more explicit, with the clear base case (returning 0
when the final character is matched) and the clever use of a large
negative value as a starting value for the non-matching case.