Fix bug in rightmost match logic
I was puzzled at the fact that the `rightmost_match_p` check in my `for`
loop here wasn't sufficient, and that the additional `haystack_len`
check was required in order to keep the "pathological" benchmark test
from regressing, indicating a bug.

Sure enough, I had clowny logic in there, causing us to not always
populate the `rightmost_match_p` array properly. For runs of characters
(like my "aaaaaaa" benchmark strings) we would end up setting the same
index for multiple characters (ie. too far to the right), preventing us
from short-circuiting adequately.

Benchmarks aren't affected much (variation is well within the bounds of
noise) because most searches aren't pathological in nature:

  Before (compute first score only):

                          user     system      total        real
  pathological        0.010000   0.000000   0.010000 (  0.008182)
  command-t           0.220000   0.000000   0.220000 (  0.222168)
  chromium (subset)   0.220000   0.020000   0.240000 (  0.144871)
  chromium (whole)    0.700000   0.010000   0.710000 (  0.442378)

  Before (compute all):

                          user     system      total        real
  pathological        0.130000   0.000000   0.130000 (  0.124744)
  command-t           0.240000   0.000000   0.240000 (  0.244833)
  chromium (subset)   0.220000   0.020000   0.240000 (  0.149158)
  chromium (whole)    0.680000   0.000000   0.680000 (  0.434695)

  After (compute first score only):

                          user     system      total        real
  pathological        0.010000   0.000000   0.010000 (  0.008652)
  command-t           0.230000   0.000000   0.230000 (  0.228463)
  chromium (subset)   0.200000   0.010000   0.210000 (  0.139945)
  chromium (whole)    0.680000   0.010000   0.690000 (  0.424456)

  After (compute all):

                          user     system      total        real
  pathological        0.110000   0.000000   0.110000 (  0.118288)
  command-t           0.230000   0.010000   0.240000 (  0.233357)
  chromium (subset)   0.210000   0.020000   0.230000 (  0.144001)
  chromium (whole)    0.650000   0.000000   0.650000 (  0.409292)

So, maybe better, and probably not worse. Still a bit disappointing that
this doesn't save us more, but I added logging to see how much sooner we
can short-circuit with this heuristic vs a length-only one. In many
cases there is zero advantage. In others we might be able to abort 4 or
10 or 20 or even more characters earlier; but in practice the memoized
algorithm is fast enough that it makes little difference.