Add inline docs explaining how the memoization arithmetic works
I originally stumbled across this compact and fast representation by
sheer luck. I don't think I full understood it until I went through and
made diagrams showing it. So, in this commit, I embed some of this in
the code, to help make clear why things are the way they are.