Rewrite using old algorithm
This is hard to explain but I am going to give it a shot.

Commit 187bc18 changed our search into a "greedy" depth-first approach,
in the sense that we recursed (continuing to pursue the in-progress
match) instead of bumping our counter and finding matches starting
further to the right. Given that we only memoize on unwinding the
recursive call stack, this means that we were prematurely locking in
some potentially sub-optimal scorings early on in the search.

Even though we later go back and exhaustively explore the problem space,
we wind up using a memoized value which is not necessarily the best
possible value for a given `needle_idx`/`haystack_idx` pair because it
is based on `distance`, and that can actually vary.

Consider this example:  a search like "matchh" and a pair of candidate
strings:

- vendor/bundle/ruby/1.8/gems/rspec-expectations-2.14.5/spec/rspec/matchers/has_spec.rb
- ruby/command-t/match.h

We score the longer "vendor" string identically to the old algorithm,
but the shorter string is scored lower, resulting in the results coming
back in the wrong order like this:

- Match first "m" in "ruby/command-t/match.h".
- Proceed, scoring with some recursion and looping, until we hit the
  last "a", memoizing its score.
- After unwinding the stack, eventually loop until the last "m" is
  matched against the "m" in the needle and scored.
- Recurse to the "a", and pull out the memoized score.
- Note that the score that got memoized for the "a" is used, but it is
  wrong, because it was computed using the distance from some other,
  previously-seen "m" and not the one we care about here.

In other words, we end up memoizing the first-seen score instead of the
best-seen score, and the reason the first-seen is not the best-seen is
that our left-biased search pattern actually tends to maximize distance.

The old algorithm didn't suffer from this problem as it is right-biased,
and by the time we unwind the stack (and commit to memoizing), we have
already seen all possible scores in the recursive subtree and can choose
the best one.

This is obvious when you think about it: the new algorithm built the
score up from the right (starting at zero for the rightmost match),
while the old algorithm built it up from the left (passing it down into
the recursive calls).

Rather than revert back to what I had originally, I've rewritten it
using my newfound insights about how the memoization works. So, we go
from "loop over haystack + recurse to explore needle" to "loop over
needle, loop over haystack + recurse to find better match".

Improvements over the original (comparing to current `master` branch
version at 4c7f02c5a) and other changes:

- No more ugly `seen_score` variable.
- `rightmost_match_p`-aware range limiting.
- No more ugly `found` variable.
- `return` in place rather than `goto` (debatable whether this is
  technically better, but it does look nicer).
- No more mutation of `haystack_idx` in the loop condition or at the end
  of the loop body.
- Simpler exit logic (no `break`).
- No more awkward `++j` preincrement.
- Remove guard at top of function and instead guard before recursing
  (makes cleaner).

This fixes our regression test but breaks a couple of others. Consider
this a work in progress. Perf seems as good or better than before. More
to come.