Remove mistaken memoization cleverness
So, I realized that my thinking here was flawed and the math I'd done to
show that this would work was flawed, and worked out magically just
because I'd chosen a rectangle with very specific dimensions. Bad luck.
Here's a counter example:

- Consider needle length 5, haystack length 100. Observe the following
  collisions with the formula:

     needle_idx * needle_len + haystack_idx =     x
          0           0            80       =     80
          1           5            75       =     80
          2           10           70       =     80
          3           15           65       =     80
          4           20           60       =     80

So with all those conditions you can expect a faster run because of all
the (false) cache hits. But I think in practice it didn't matter much at
all because the matrix is still sparse enough that hits were pretty
rare. Either that, or I had a compensating bug somewhere else. Some
support is leant to theory that collisions were rare, or the impact of
them on scoring was relatively low, because I wasn't aware of any
noticeable ranking bugs.

  Before [ie. 1f41e7630]  (best of three runs):

                            user     system      total        real
    pathological        2.920000   0.010000   2.930000 (  2.927735)
    command-t           1.850000   0.010000   1.860000 (  1.854239)
    chromium (subset)   2.260000   0.200000   2.460000 (  1.510716)
    chromium (whole)   11.090000   0.040000  11.130000 (  6.670843)

  After (best of three runs):

                            user     system      total        real
    pathological        2.170000   0.010000   2.180000 (  2.185652)
    command-t           1.750000   0.010000   1.760000 (  1.753948)
    chromium (subset)   2.270000   0.200000   2.470000 (  1.535449)
    chromium (whole)   10.940000   0.040000  10.980000 (  6.708795)

Note that for the "before" test I went back to 1f41e7630 to get past the
intermediate broken states and weirdness that I've had on this branch of
late.