Add heap implementation
This is a pretty standard binary min heap[*], basically based on the JS one
that I wrote for Corpus:

https://github.com/wincent/corpus/blob/next/src/app/Heap.js

[*]: Except for the fact that this one is fixed size for reasons which I
am about to explain.

Profiling shows that we're spending about half our time in `qsort` which
is `O(n log n)` where `n` is the number of files. The hypothesis that I
want to test out here is that we might be able to do less work here if
we use a fixed size heap as follows:

- `n` files, `m` threads, and a desire to show the top `l` matches.
- One heap per thread to avoid locking overhead.
- `l` slots per heap.
- Insert first `l` results in bulk mode (`O(l)`).
- From then on, do a normal insert for each result (`O(n log l)`).
- Note that `l` is dwarfed by `n`, so this should be a lot less work.
- After each insertion, extract the minimum value (`O(n log l)`); this
  means our "min" heap actually serves as an always up-to-date view of
  the "max" scoring matches.
- Also note that we can avoid doing the insert at all by peeking at the
  root of the heap and seeing if the score is already higher than what
  we would insert.
- At end, extract all; or better still, just read straight out (`O(ml)`
  instead of `O(ml log l)`).
- Sort that (`O(ml log ml`).

I believe this could knock off anywhere from 5 to 80% of our sorting
time, which is itself about 50% of the overall time. Of course, there
are lots of reasons why this could be wrong (including the fact that
there is some overhead in the heap management code that might make it
less competitive with the highly optimized standard library `qsort`).
None the less, this is a hypothesis worth checking out.

Crude numbers to show my working for that 5-80% guess:

- n = 500,000; m = 5; l = 32
- sorting =
    n log n comparisons =
    500,000 log 500,000 =
    500,000 * 19 =
    9.5 million comparisons
- heap approach =
    m * l bulk inserts (2,500 operations) +
    n/5 * 5 * log 32 (ie. 5) inserts (500,000 operations) +
    same for extractions (500,000 operations) +
    m * l to read out (2,500 operations) +
    250 log 250 (ie. 8) to sort (2,000 operations) =
    1 million operations

Unknowns:

- What are the relative costs of the operations?
- How many of the 1 million insert + extract pairs can we eliminate by
  peeking at the minimum score, and to what benefit?
- How accurate was my profiling that suggests that half our time is
  spent sorting?

As an initial test, I turned off sorting in the benchmarks to see the
cost and saw them drop from 7.4s for the Chromium set down to 0.95s. I
then turned sorting back on and added heap insert + delete and saw them
go up 0.5s. In other words, this could end up dropping us from 7.4s to
around 1.5s, which would be amazing.

As for the heap code, I haven't tested it yet, but it does compile
without errors.

> Beware of bugs in the above code; I have only proved it correct, not
> tried it.

(Knuth)