Align limit default with CommandTMaxHeight default
In Vim, limit is always going to be 15 (unless overridden by the user)
or possibly less if the window height does not permit it. Even if they
set the height to "unlimited" it will still be clipped to the window
height.

As such, to make benchmarking more closely model editor-based usage
scenarios, I'm tweaking the low-level default too.

This commit means we're doing a little less work (fewer `Array#push`
ops), but there is not much effect:

  Before (best of three):

                            user     system      total        real
    pathological        0.250000   0.010000   0.260000 (  0.253167)
    command-t           1.710000   0.000000   1.710000 (  1.722218)
    chromium (subset)   2.160000   0.210000   2.370000 (  1.511613)
    chromium (whole)   10.370000   0.040000  10.410000 (  6.640190)

  After (best of three):

                            user     system      total        real
    pathological        0.250000   0.000000   0.250000 (  0.252433)
    command-t           1.970000   0.020000   1.990000 (  1.985316)
    chromium (subset)   2.090000   0.200000   2.290000 (  1.450036)
    chromium (whole)   10.290000   0.040000  10.330000 (  6.495397)

Note that we haven't actually made anything faster here; we've just made
the benchmark better align with real world usage, like we did in
fb1ab6b.