Change heap overflow strategy
I think this is a less leaky abstraction. If you want the
insert-then-extract behavior, you can request a one-larger heap
yourself.

In this way, callers don't need to know the internal implementation
details of the heap to predict which node will be evicted on overflow.
The behavior is much more easily defined (excess nodes are outright
ignored).