Update optimization info in docs
These numbers are doomed to go stale before I release, but anyway, I
re-measured all the way back to 1.6 (the first version to use
parallelized search):

  next:

                          user     system      total        real
  pathological        0.200000   0.000000   0.200000 (  0.197547)
  command-t           1.250000   0.010000   1.260000 (  1.265484)
  chromium (subset)   0.960000   0.130000   1.090000 (  0.217825)
  chromium (whole)    3.850000   0.020000   3.870000 (  0.533365)

  3.0.2 (current released version):

                          user     system      total        real
  pathological        2.580000   0.010000   2.590000 (  2.589704)
  command-t           2.070000   0.000000   2.070000 (  2.079982)
  chromium (subset)   3.650000   0.180000   3.830000 (  1.790363)
  chromium (whole)   24.210000   0.060000  24.270000 (  9.096863)

  1.6 (first version to use parallelized search):

                          user     system      total        real
  pathological        0.070000   0.000000   0.070000 (  0.070763)
  command-t           2.260000   0.010000   2.270000 (  2.271704)
  chromium (subset)   9.700000   0.170000   9.870000 (  3.013519)
  chromium (whole)   51.960000   0.070000  52.030000 ( 13.503562)

So, in the 3+ years between 1.6 and 3.0.2 the implementation got quite a
bit faster, about 33% faster, even as it also got improvements to
improve its correctness (at the cost of speed). But current next blows
the doors right off, at about 17x faster than 3.0.2.

Still want to do some refactoring, testing and baking before releasing
though, so changed the anticipated release date in the docs to March
(instead of February, of which there is only one day left). Also made
the description of the optimizations a little more vague, because there
are many of them.